@@ -4,11 +4,11 @@ version: 1.1.4
 homepage: https://github.com/InaldoManso/Simple_Player
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
-  video_player: ">2.4.6 <3.0.0"
+  video_player: ">=2.8.2 <=3.0.0"
   screen_brightness: ">=0.2.2 <1.0.0"
   flutter:
     sdk: flutter
