@@ -9,15 +9,17 @@ import '../constants/constants.dart';
 import 'dart:async';
 
 class SimplePlayerFullScreen extends StatefulWidget {
-  final SimpleController simpleController;
-  final SimplePlayerSettings simplePlayerSettings;
-  final SimplePlayerState simplePlayerState;
-  const SimplePlayerFullScreen(
-      {Key? key,
-      required this.simpleController,
-      required this.simplePlayerSettings,
-      required this.simplePlayerState})
-      : super(key: key);
+  final VideoPlayerController videoPlayerController;
+  // final SimpleController simpleController;
+  // final SimplePlayerSettings simplePlayerSettings;
+  // final SimplePlayerState simplePlayerState;
+  const SimplePlayerFullScreen({
+    Key? key,
+    required this.videoPlayerController,
+    // required this.simpleController,
+    // required this.simplePlayerSettings,
+    // required this.simplePlayerState,
+  }) : super(key: key);
 
   @override
   State<SimplePlayerFullScreen> createState() => _SimplePlayerFullScreenState();
@@ -120,20 +122,20 @@ class _SimplePlayerFullScreenState extends State<SimplePlayerFullScreen>
 
   /// Retrieves and inserts the last state of the previous screen
   _lastState() {
-    SimplePlayerState simplePlayerState = widget.simplePlayerState;
+    // SimplePlayerState simplePlayerState = widget.simplePlayerState;
 
-    setState(() {
-      _speed = simplePlayerState.speed;
-      _tittle = simplePlayerState.label;
-      _wasPlaying = simplePlayerState.wasPlaying;
-      _confortMode = simplePlayerState.confortMode;
-    });
+    //   setState(() {
+    //     _speed = simplePlayerState.speed;
+    //     _tittle = simplePlayerState.label;
+    //     _wasPlaying = simplePlayerState.wasPlaying;
+    //     _confortMode = simplePlayerState.confortMode;
+    //   });
 
-    /// Methods
-    _videoPlayerController.setLooping(simplePlayerState.loopMode!);
-    _jumpTo(simplePlayerState.currentSeconds!);
-    _speedSetter(simplePlayerState.speed);
-    _configureRotation();
+    //   /// Methods
+    //   _videoPlayerController.setLooping(simplePlayerState.loopMode!);
+    //   _jumpTo(simplePlayerState.currentSeconds!);
+    //   _speedSetter(simplePlayerState.speed);
+    //   _configureRotation();
   }
 
   _configureRotation() {
@@ -219,53 +221,66 @@ class _SimplePlayerFullScreenState extends State<SimplePlayerFullScreen>
 
   /// Update the real-time seconds counter on replay.
   _secondsListener() {
-    _videoPlayerController.addListener(
-      () {
-        widget.simpleController.updateController(_videoPlayerController);
-        bool playing = _videoPlayerController.value.isPlaying;
-        if (_currentSeconds == _totalSeconds && !playing) {
-          _showAndHideControls(true);
-          _animationController.reverse();
-          _jumpTo(0.0);
-        }
-        setState(() {
-          _currentSeconds =
-              _videoPlayerController.value.position.inMilliseconds.toDouble();
-          _showTime = DateFormatter()
-              .currentTime(_videoPlayerController.value.position);
-        });
-      },
-    );
+    // _videoPlayerController.addListener(
+    //   () {
+    //     widget.simpleController.updateController(_videoPlayerController);
+    //     bool playing = _videoPlayerController.value.isPlaying;
+    //     if (_currentSeconds == _totalSeconds && !playing) {
+    //       _showAndHideControls(true);
+    //       _animationController.reverse();
+    //       _jumpTo(0.0);
+    //     }
+    //     setState(() {
+    //       _currentSeconds =
+    //           _videoPlayerController.value.position.inMilliseconds.toDouble();
+    //       _showTime = DateFormatter()
+    //           .currentTime(_videoPlayerController.value.position);
+    //     });
+    //   },
+    // );
   }
 
   /// Shows or hides the HUB from controller commands.
   _listenerPlayFromController() {
-    String changeTime = '';
-    widget.simpleController.listenPlayAndPause().listen((event) {
-      if (changeTime != event) {
-        bool playing = _videoPlayerController.value.isPlaying;
-        if (playing) {
-          _showAndHideControls(false);
-        } else {
-          _showAndHideControls(true);
-        }
-        changeTime = event;
-      }
-    });
+    // String changeTime = '';
+    // widget.simpleController.listenPlayAndPause().listen((event) {
+    //   if (changeTime != event) {
+    //     bool playing = _videoPlayerController.value.isPlaying;
+    //     if (playing) {
+    //       _showAndHideControls(false);
+    //     } else {
+    //       _showAndHideControls(true);
+    //     }
+    //     changeTime = event;
+    //   }
+    // });
   }
 
   /// Responsible for starting the interface
   _initializeInterface() {
-    simplePlayerSettings = widget.simplePlayerSettings;
+    // simplePlayerSettings = widget.simplePlayerSettings;
 
-    setState(() {
-      _colorAccent = simplePlayerSettings.colorAccent;
+    // setState(() {
+    //   _colorAccent = simplePlayerSettings.colorAccent;
+    // });
+
+    // /// Methods
+    // _setupControllers(simplePlayerSettings);
+    // _secondsListener();
+    // _listenerPlayFromController();
+
+    print('PPX Play em 3 segundos');
+    Timer(Duration(seconds: 3), () {
+      print('PPX Play');
+      widget.videoPlayerController.play();
+
+      // ==============================
+      print('PPX Pause em 1 segundo');
+      Timer(Duration(seconds: 1), () {
+        print('PPX Pause');
+        widget.videoPlayerController.pause();
+      });
     });
-
-    /// Methods
-    _setupControllers(simplePlayerSettings);
-    _secondsListener();
-    _listenerPlayFromController();
   }
 
   @override
@@ -276,200 +291,208 @@ class _SimplePlayerFullScreenState extends State<SimplePlayerFullScreen>
     super.initState();
   }
 
-  @override
-  void dispose() {
-    _dismissConstrollers();
-    super.dispose();
-  }
+  // @override
+  // void dispose() {
+  //   _dismissConstrollers();
+  //   super.dispose();
+  // }
 
-  /// Finalize resources
-  _dismissConstrollers() async {
-    _animationController.stop();
-    _animationController.dispose();
-    _videoPlayerController.removeListener(() {});
-    _videoPlayerController.dispose();
-  }
+  // /// Finalize resources
+  // _dismissConstrollers() async {
+  //   _animationController.stop();
+  //   _animationController.dispose();
+  //   _videoPlayerController.removeListener(() {});
+  //   _videoPlayerController.dispose();
+  // }
 
   @override
   Widget build(BuildContext context) {
-    double height = MediaQuery.of(context).size.height;
-    double width = MediaQuery.of(context).size.width;
-    return WillPopScope(
-      onWillPop: () async {
-        _fullScreenManager();
-        return false;
-      },
-      child: Material(
-        color: Colors.black,
-        child: _playbackSetUp!
-            ? Container(
-                width: width,
-                color: Colors.black,
-                child: Stack(
-                  children: [
-                    Center(
-                      child: AspectRatio(
-                        aspectRatio:
-                            _aspectRatioManager(_videoPlayerController),
-                        child: VideoPlayer(_videoPlayerController),
-                      ),
-                    ),
-                    _videoPlayerController.value.isInitialized
-                        ? Center(
-                            child: AnimatedSwitcher(
-                              duration: const Duration(milliseconds: 200),
-                              transitionBuilder:
-                                  (Widget child, Animation<double> animation) {
-                                return FadeTransition(
-                                    opacity: animation, child: child);
-                              },
-                              child: _visibleControls!
-                                  ? GestureDetector(
-                                      child: AnimatedContainer(
-                                        duration: const Duration(seconds: 1),
-                                        key: const ValueKey('a'),
-                                        color: _confortMode!
-                                            ? Colors.deepOrange.withOpacity(0.1)
-                                            : Colors.black.withOpacity(0.2),
-                                        child: Column(
-                                          crossAxisAlignment:
-                                              CrossAxisAlignment.stretch,
-                                          children: [
-                                            Row(
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment
-                                                      .spaceBetween,
-                                              children: [
-                                                Padding(
-                                                  padding:
-                                                      const EdgeInsets.only(
-                                                          left: 16),
-                                                  child: Text(
-                                                    _tittle!,
-                                                    style: const TextStyle(
-                                                      color: Colors.white,
-                                                      fontSize: 16,
-                                                      shadows: [
-                                                        Shadow(
-                                                          blurRadius: 10.0,
-                                                          color: Colors.black,
-                                                          offset:
-                                                              Offset(3.0, 2.0),
-                                                        ),
-                                                      ],
-                                                    ),
-                                                  ),
-                                                ),
-                                                IconButton(
-                                                  padding:
-                                                      const EdgeInsets.all(0),
-                                                  icon: const Icon(
-                                                    Icons.settings_outlined,
-                                                    color: Colors.white,
-                                                  ),
-                                                  onPressed: () {
-                                                    _showScreenSettings();
-                                                  },
-                                                ),
-                                              ],
-                                            ),
-                                            Expanded(
-                                              child: Center(
-                                                child: IconButton(
-                                                    icon: AnimatedIcon(
-                                                        size: 40,
-                                                        color: Colors.white,
-                                                        icon: AnimatedIcons
-                                                            .play_pause,
-                                                        progress:
-                                                            _animationController),
-                                                    onPressed: () =>
-                                                        _playAndPauseSwitch(
-                                                            pauseButton: true)),
-                                              ),
-                                            ),
-                                            Row(
-                                              mainAxisSize: MainAxisSize.min,
-                                              children: [
-                                                Padding(
-                                                    padding:
-                                                        const EdgeInsets.only(
-                                                            left: 16),
-                                                    child: Text(_showTime!,
-                                                        style: const TextStyle(
-                                                            color:
-                                                                Colors.white))),
-                                                Expanded(
-                                                    child: SliderTheme(
-                                                  data: constants
-                                                      .getSliderThemeData(
-                                                          colorAccent:
-                                                              _colorAccent),
-                                                  child: Slider.adaptive(
-                                                    value: _currentSeconds!,
-                                                    max: _totalSeconds!,
-                                                    min: 0,
-                                                    label: _currentSeconds
-                                                        .toString(),
-                                                    onChanged: (double value) {
-                                                      _jumpTo(value);
-                                                    },
-                                                  ),
-                                                )),
-                                                IconButton(
-                                                  padding:
-                                                      const EdgeInsets.all(0),
-                                                  icon: const Icon(
-                                                    Icons.fullscreen_exit,
-                                                    color: Colors.white,
-                                                  ),
-                                                  onPressed: () =>
-                                                      _fullScreenManager(),
-                                                ),
-                                              ],
-                                            ),
-                                          ],
-                                        ),
-                                      ),
-                                      onTap: () => _screenTap(),
-                                    )
-                                  : GestureDetector(
-                                      child: AnimatedContainer(
-                                        duration: const Duration(seconds: 1),
-                                        key: const ValueKey('b'),
-                                        color: _confortMode!
-                                            ? Colors.deepOrange.withOpacity(0.1)
-                                            : Colors.transparent,
-                                        height: height,
-                                      ),
-                                      onTap: () => _screenTap(),
-                                    ),
-                            ),
-                          )
-                        : const Center(child: CircularProgressIndicator()),
-                    AnimatedSwitcher(
-                      duration: const Duration(milliseconds: 200),
-                      transitionBuilder:
-                          (Widget child, Animation<double> animation) {
-                        return FadeTransition(opacity: animation, child: child);
-                      },
-                      child: _visibleSettings!
-                          ? SettingsScreen(
-                              colorAccent: _colorAccent!,
-                              speed: _speed!,
-                              confortModeOn: _confortMode!,
-                              onExit: () => _showScreenSettings(),
-                              confortClicked: (value) =>
-                                  setState(() => _confortMode = value),
-                              speedSelected: (value) => _speedSetter(value),
-                            )
-                          : const SizedBox(width: 1, height: 1),
-                    ),
-                  ],
-                ),
-              )
-            : Center(child: CircularProgressIndicator()),
+    return Scaffold(
+      body: GestureDetector(
+        onTap: () {
+          Navigator.pop(context);
+        },
+        child: VideoPlayer(widget.videoPlayerController),
       ),
     );
+    // double height = MediaQuery.of(context).size.height;
+    // double width = MediaQuery.of(context).size.width;
+    // return WillPopScope(
+    //   onWillPop: () async {
+    //     _fullScreenManager();
+    //     return false;
+    //   },
+    //   child: Material(
+    //     color: Colors.black,
+    //     child: _playbackSetUp!
+    //         ? Container(
+    //             width: width,
+    //             color: Colors.black,
+    //             child: Stack(
+    //               children: [
+    //                 Center(
+    //                   child: AspectRatio(
+    //                     aspectRatio:
+    //                         _aspectRatioManager(_videoPlayerController),
+    //                     child: VideoPlayer(_videoPlayerController),
+    //                   ),
+    //                 ),
+    // _videoPlayerController.value.isInitialized
+    //     ? Center(
+    //         child: AnimatedSwitcher(
+    //           duration: const Duration(milliseconds: 200),
+    //           transitionBuilder:
+    //               (Widget child, Animation<double> animation) {
+    //             return FadeTransition(
+    //                 opacity: animation, child: child);
+    //           },
+    //           child: _visibleControls!
+    //               ? GestureDetector(
+    //                   child: AnimatedContainer(
+    //                     duration: const Duration(seconds: 1),
+    //                     key: const ValueKey('a'),
+    //                     color: _confortMode!
+    //                         ? Colors.deepOrange.withOpacity(0.1)
+    //                         : Colors.black.withOpacity(0.2),
+    //                     child: Column(
+    //                       crossAxisAlignment:
+    //                           CrossAxisAlignment.stretch,
+    //                       children: [
+    //                         Row(
+    //                           mainAxisAlignment:
+    //                               MainAxisAlignment
+    //                                   .spaceBetween,
+    //                           children: [
+    //                             Padding(
+    //                               padding:
+    //                                   const EdgeInsets.only(
+    //                                       left: 16),
+    //                               child: Text(
+    //                                 _tittle!,
+    //                                 style: const TextStyle(
+    //                                   color: Colors.white,
+    //                                   fontSize: 16,
+    //                                   shadows: [
+    //                                     Shadow(
+    //                                       blurRadius: 10.0,
+    //                                       color: Colors.black,
+    //                                       offset:
+    //                                           Offset(3.0, 2.0),
+    //                                     ),
+    //                                   ],
+    //                                 ),
+    //                               ),
+    //                             ),
+    //                             IconButton(
+    //                               padding:
+    //                                   const EdgeInsets.all(0),
+    //                               icon: const Icon(
+    //                                 Icons.settings_outlined,
+    //                                 color: Colors.white,
+    //                               ),
+    //                               onPressed: () {
+    //                                 _showScreenSettings();
+    //                               },
+    //                             ),
+    //                           ],
+    //                         ),
+    //                         Expanded(
+    //                           child: Center(
+    //                             child: IconButton(
+    //                                 icon: AnimatedIcon(
+    //                                     size: 40,
+    //                                     color: Colors.white,
+    //                                     icon: AnimatedIcons
+    //                                         .play_pause,
+    //                                     progress:
+    //                                         _animationController),
+    //                                 onPressed: () =>
+    //                                     _playAndPauseSwitch(
+    //                                         pauseButton: true)),
+    //                           ),
+    //                         ),
+    //                         Row(
+    //                           mainAxisSize: MainAxisSize.min,
+    //                           children: [
+    //                             Padding(
+    //                                 padding:
+    //                                     const EdgeInsets.only(
+    //                                         left: 16),
+    //                                 child: Text(_showTime!,
+    //                                     style: const TextStyle(
+    //                                         color:
+    //                                             Colors.white))),
+    //                             Expanded(
+    //                                 child: SliderTheme(
+    //                               data: constants
+    //                                   .getSliderThemeData(
+    //                                       colorAccent:
+    //                                           _colorAccent),
+    //                               child: Slider.adaptive(
+    //                                 value: _currentSeconds!,
+    //                                 max: _totalSeconds!,
+    //                                 min: 0,
+    //                                 label: _currentSeconds
+    //                                     .toString(),
+    //                                 onChanged: (double value) {
+    //                                   _jumpTo(value);
+    //                                 },
+    //                               ),
+    //                             )),
+    //                             IconButton(
+    //                               padding:
+    //                                   const EdgeInsets.all(0),
+    //                               icon: const Icon(
+    //                                 Icons.fullscreen_exit,
+    //                                 color: Colors.white,
+    //                               ),
+    //                               onPressed: () =>
+    //                                   _fullScreenManager(),
+    //                             ),
+    //                           ],
+    //                         ),
+    //                       ],
+    //                     ),
+    //                   ),
+    //                   onTap: () => _screenTap(),
+    //                 )
+    //               : GestureDetector(
+    //                   child: AnimatedContainer(
+    //                     duration: const Duration(seconds: 1),
+    //                     key: const ValueKey('b'),
+    //                     color: _confortMode!
+    //                         ? Colors.deepOrange.withOpacity(0.1)
+    //                         : Colors.transparent,
+    //                     height: height,
+    //                   ),
+    //                   onTap: () => _screenTap(),
+    //                 ),
+    //         ),
+    //       )
+    //     : const Center(child: CircularProgressIndicator()),
+    // AnimatedSwitcher(
+    //   duration: const Duration(milliseconds: 200),
+    //   transitionBuilder:
+    //       (Widget child, Animation<double> animation) {
+    //     return FadeTransition(opacity: animation, child: child);
+    //   },
+    //   child: _visibleSettings!
+    //       ? SettingsScreen(
+    //           colorAccent: _colorAccent!,
+    //           speed: _speed!,
+    //           confortModeOn: _confortMode!,
+    //           onExit: () => _showScreenSettings(),
+    //           confortClicked: (value) =>
+    //               setState(() => _confortMode = value),
+    //           speedSelected: (value) => _speedSetter(value),
+    //         )
+    //       : const SizedBox(width: 1, height: 1),
+    // ),
+    //               ],
+    //             ),
+    //           )
+    //         : Center(child: CircularProgressIndicator()),
+    //   ),
+    // );
   }
 }
