@@ -2,8 +2,10 @@ library simple_player;
 
 /// Packages
 import 'package:simple_player/aplication/simple_controller.dart';
+import 'package:simple_player/constants/constants.dart';
 import 'package:simple_player/presentation/simple_player_screen.dart';
 import 'package:simple_player/model/simple_player_settings.dart';
+import 'package:video_player/video_player.dart';
 import 'package:flutter/material.dart';
 
 /// Resources
@@ -54,22 +56,69 @@ class SimplePlayer extends StatefulWidget {
   ///Keep an eye out for examples. 🕶️
   ///Good coding! 😎💙
   ///
-  const SimplePlayer(
-      {Key? key,
-      required this.simpleController,
-      required this.simplePlayerSettings})
-      : super(key: key);
+  const SimplePlayer({
+    Key? key,
+    required this.simpleController,
+    required this.simplePlayerSettings,
+  }) : super(key: key);
 
   @override
   State<SimplePlayer> createState() => _SimplePlayerState();
 }
 
 class _SimplePlayerState extends State<SimplePlayer> {
+  // Attributes
+  late VideoPlayerController videoPlayerController;
+  bool loaded = false;
+
+  getControler(SimplePlayerSettings simplePlayerSettings) async {
+    switch (simplePlayerSettings.type) {
+      case 'network':
+        videoPlayerController = VideoPlayerController.networkUrl(
+            Uri.parse(simplePlayerSettings.path));
+        break;
+
+      case 'assets':
+        videoPlayerController =
+            VideoPlayerController.asset(simplePlayerSettings.path);
+        break;
+
+      default:
+        videoPlayerController =
+            VideoPlayerController.networkUrl(Uri.parse(Constants.videoExample));
+        break;
+    }
+
+    // Ready to play
+    await videoPlayerController.initialize();
+    setState(() => loaded = true);
+  }
+
+  @override
+  void initState() {
+    getControler(widget.simplePlayerSettings);
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return SimplePlayerScrren(
-      simpleController: widget.simpleController,
-      simplePlayerSettings: widget.simplePlayerSettings,
-    );
+    if (loaded) {
+      return SimplePlayerScrren(
+        videoPlayerController: videoPlayerController,
+        simpleController: widget.simpleController,
+        simplePlayerSettings: widget.simplePlayerSettings,
+      );
+    } else {
+      return AspectRatio(
+        aspectRatio: widget.simplePlayerSettings.aspectRatio,
+        child: Container(
+          color: Colors.black,
+          alignment: Alignment.center,
+          child: CircularProgressIndicator(
+            color: widget.simplePlayerSettings.colorAccent,
+          ),
+        ),
+      );
+    }
   }
 }
